+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 -  export to the stdout
+-  import from stdin
 
 ### Removed
 -  bitwarden support